+++ src/main/java/com/jmmacedo/calculator/application/rest/adapters/input/Operation.java
@@ -1,0 +1,6 @@
+package com.jmmacedo.calculator.application.rest.adapters.input;
+
+public interface Operation {
+
+  int operate(String operation, Integer baseNumber, Integer number);
+}