+++ CHANGELOG.md
@@ -2,3 +2,8 @@
 
 Add controller
 Add interceptor
+
+# Add Custom Exception and Exception Handler
+
+- Add a custom exception
+- Add an exception handler with controller advice annotation